+++ Release Notes.md
@@ -25,6 +25,7 @@ SOFTWARE.
 
 * moved defines & inlines from arm_hal_peripheral to arm_family headers
 * arm_macros: added ROOT_OF macro
+* arm_macros: added L qualifier for LSHIFT & RSHIFT (compatibility with less than 32b CPUs)
 * arm_inlines: const qualifier for most inline parameters
 * arm_inlines: modified TPSSUP_MS & TPSINF_MS to handle custom tick frequency (following arm_family header or custom defines)
 * arm_chip_sam: refactored doxygen comments block